Extract out plist processing as an extension and invoke via 'Edit display format'. Compiles but gets stuck when 'plist' is selcted, however selecting a different format clears the issue.
