+++ src/EditDialog.cpp
@@ -21,591 +21,6 @@
 #include <QPainter>
 #include <QClipboard>
 
-//  The following section of code should really be in a separate class
-//  but this is only a test!
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <ctype.h>
-#include <time.h>
-
-#define INCREMENT 4
-
-#define ERROR_NONE                   0
-#define ERROR_INSUFFICIENT_MEMORY    1
-#define ERROR_INVALID_HEADER         2
-#define ERROR_INVALID_TRAILER        3
-#define ERROR_INVALID_OFFSET         4
-#define ERROR_INVALID_OBJECT_LENGTH  5
-#define ERROR_INVALID_REFERENCE      6
-
-typedef struct KEY {
-  struct OBJECT *key;
-  struct OBJECT *value;
-} KEY;
-
-typedef struct OBJECT {
-  int type;
-  int length;
-  union {
-    time_t date;
-    double real;
-    long integer;
-    unsigned long uid;
-    unsigned char binary[1];
-    char text[1];
-    short int utf16[1];
-    int refs[1];
-    KEY keys[1];
-    struct OBJECT *objects[1];
-  } data;
-} OBJECT;
-
-typedef struct CONFIG {
-  int bytesPerOffset;
-  int bytesPerReference;
-  int objectCount;
-  int rootObjectReference;
-  int offsetTableOffset;
-  size_t bufferLength;
-  size_t outputBufferLength;
-  size_t outputBufferIn;
-  long *offsetTable;
-  unsigned char *buffer;
-  unsigned char *outputBuffer;
-} CONFIG;
-
-int createObject(int type, int length, int extra, OBJECT **ptr2);
-int readObject(CONFIG *cfg, long offset, OBJECT **ptr2);
-
-static int indent = 2;
-
-int outputText(CONFIG *cfg, const char *text)
-{
-  size_t textLength = strlen((const char *)text);
-  size_t availableSpace = cfg->outputBufferLength - cfg->outputBufferIn;
-  while (textLength >= availableSpace) {
-    cfg->outputBufferLength += textLength + 1024;
-    cfg->outputBuffer = (unsigned char *)realloc(cfg->outputBuffer, cfg->outputBufferLength);
-    if (cfg->outputBuffer == NULL)
-      return 1;
-    availableSpace = cfg->outputBufferLength - cfg->outputBufferIn;
-  }
-  strcpy((char *)(cfg->outputBuffer+cfg->outputBufferIn), text);
-  cfg->outputBufferIn += textLength;
-  return ERROR_NONE;
-}
-
-int printWithIndent(CONFIG *cfg, const char *text, int newline)
-{
-  int err = ERROR_NONE;
-  char spaces[9] = "        ";
-  int n = indent;
-  while ((n > 8) && (err == ERROR_NONE)) {
-    err = outputText(cfg, spaces);
-    n -= 8;
-  }
-  if ((n > 0) && (err == ERROR_NONE))
-    err = outputText(cfg, spaces + 8 - n);
-  if (err == ERROR_NONE)
-    err = outputText(cfg, text);
-  if (newline && (err == ERROR_NONE))
-    err = outputText(cfg, "\n");
-  return err;
-}
-
-int readHeader(CONFIG *cfg)
-{
-  if ((cfg->bufferLength > 40) &&
-     ((strncmp((const char *)(cfg->buffer), "bplist0", 7) != 0) ||
-     ((cfg->buffer[7] != '0') && (cfg->buffer[7] != '1'))))
-    return ERROR_INVALID_HEADER;
-  return ERROR_NONE;
-}
-
-int readTrailer(CONFIG *cfg)
-{
-  int i, j;
-  int objectCount = 0;
-  int rootObjectReference = 0;
-  int offsetTableOffset = 0;
-  unsigned char *ptr;
-  unsigned char *buffer = cfg->buffer + cfg->bufferLength - 32;
-
-  //  Extract the relevant fields
-  for (i=12; i < 16; i++)
-    objectCount = (objectCount << 8) + buffer[i];
-  for (i=20; i < 24; i++)
-    rootObjectReference = (rootObjectReference << 8) + buffer[i];
-  for (i=28; i < 32; i++)
-    offsetTableOffset = (offsetTableOffset << 8) + buffer[i];
-   
-  //  Populate the configurartion structure
-  cfg->bytesPerOffset = buffer[6];
-  cfg->bytesPerReference = buffer[7];
-  cfg->objectCount = objectCount;
-  cfg->rootObjectReference = rootObjectReference;
-  cfg->offsetTableOffset = offsetTableOffset;
-  cfg->offsetTable = (long *)malloc((size_t)objectCount * sizeof(long));
-  if (cfg->offsetTable == NULL)
-    return ERROR_INSUFFICIENT_MEMORY;
-  ptr = cfg->buffer + offsetTableOffset;
-  for (i=0; i < objectCount; i++) {
-    long n = 0;
-    for (j=0; j < cfg->bytesPerOffset; j++)
-       n = (n << 8) | *(ptr++);
-    cfg->offsetTable[i] = n;
-  }
-  return ERROR_NONE;
-}
-  
-int createObject(int type, int length, int extra, OBJECT **ptr2)
-{
-  *ptr2 = NULL;
-  OBJECT *ptr = (OBJECT *)malloc(sizeof(OBJECT) + (size_t)extra);
-  if (ptr == NULL)
-    return ERROR_INSUFFICIENT_MEMORY;
-  ptr->type = type;
-  ptr->length = length;
-  *ptr2 = ptr;
-  return ERROR_NONE;
-}
-
-long readInteger(unsigned char *ptr, int desc)
-{
-  long value = 0L;
-  int n = 1 << (desc & 0x03);
-  while (n--)
-    value = (value << 8) | (long)(*(ptr++));
-  return value;
-}
-
-unsigned long readUid(unsigned char *ptr, int desc)
-{
-  unsigned long value = 0L;
-  int n = 1 << (desc & 0x03);
-  while (n--)
-    value = (value << 8) | (unsigned long)(*(ptr++));
-  return value;
-}
-
-double readDouble(unsigned char *ptr, int desc)
-{
-  union {
-    double v;
-    float f[2];
-    unsigned char b[8];
-  } value;
-  int n = 1 << (desc & 0x03);
-  for (int i=0; i < n; i++)
-    value.b[7-i] = *(ptr++);
-  if (n == 4)
-    value.v = (double)(value.f[1]);
-  return value.v;
-}
-
-int readArray(CONFIG *cfg, unsigned char *ptr, int type, int length, OBJECT **array2)
-{
-  int i;
-  long offset;
-  OBJECT *array;
-  int err = createObject(type, length, ((length-1) * (int)sizeof(OBJECT)), &array);
-  *array2 = NULL;
-  for (i=0; (i < length) && (err == ERROR_NONE); i++) {
-    offset = 0L;
-    for (int j=0; j < cfg->bytesPerReference; j++)
-      offset = (offset << 8) | (long)(*(ptr++));
-    if (offset >= cfg->objectCount)
-      return ERROR_INVALID_REFERENCE;
-    offset = cfg->offsetTable[offset];
-    err = readObject(cfg, offset, &(array->data.objects[i]));
-  }
-  if (err == ERROR_NONE)
-    *array2 = array;
-  return err;
-}
-
-int readDictionary(CONFIG *cfg, unsigned char *ptr, int type, int length, OBJECT **dict2)
-{
-  int i;
-  long offset;
-  OBJECT *dict;
-  int err  = createObject(type, length, ((length-1) * (int)sizeof(KEY)), &dict);
-  *dict2 = NULL;
-  if (err != ERROR_NONE)
-    return err;
-  for (i=0; (i < length) && (err == ERROR_NONE); i++) {
-    offset = 0L;
-    for (int j=0; j < cfg->bytesPerReference; j++)
-      offset = (offset << 8) | (long)(*(ptr++));
-    if (offset >= cfg->objectCount)
-      return ERROR_INVALID_REFERENCE;
-    offset = cfg->offsetTable[offset];
-    err = readObject(cfg, offset, &(dict->data.keys[i].key));
-  }
-  for (i=0; (i < length) && (err == ERROR_NONE); i++) {
-    offset = 0L;
-    for (int j=0; j < cfg->bytesPerReference; j++)
-      offset = (offset << 8) | (long)(*(ptr++));
-    if (offset >= cfg->objectCount)
-      return ERROR_INVALID_REFERENCE;
-    offset = cfg->offsetTable[offset];
-    err = readObject(cfg, offset, &(dict->data.keys[i].value));
-  }
-  if (err == ERROR_NONE)
-    *dict2 = dict;
-  return ERROR_NONE;
-}
-
-int readObject(CONFIG *cfg, long offset, OBJECT **ptr2)
-{
-  int i;
-  int length;
-  int type;
-  int err = ERROR_NONE;
-  OBJECT *obj;
-  unsigned char *ptr;
-
-  *ptr2 = NULL;
-  if ((size_t)offset >= cfg->bufferLength)
-    return ERROR_INVALID_OFFSET;
-  ptr = cfg->buffer + offset;
-  type = *(ptr++);
-  length = (type & 0x0F);
-  type >>= 4;
-  if ((type != 0) && (length == 0x0F)) {
-    length = *(ptr++);
-    if ((length >> 4) != 0x01)
-      return ERROR_INVALID_OBJECT_LENGTH;
-    i = (int)readInteger(ptr, length);
-    ptr += (1 << (length & 0x03));
-    length = i;
-  }
-  switch (type) {
-    case 0x00:               //  Singleton
-              err = createObject(type, length, 5, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              switch(length) {
-                case 0:             //  NULL
-                       strcpy(obj->data.text, "<Null/>");
-                       break;
-                case 8:             //  False
-                       strcpy(obj->data.text, "<False/>");
-                       break;
-                case 9:             //  True
-                       strcpy(obj->data.text, "<True/>");
-                       break;
-                case 15:             //  Fill
-                       strcpy(obj->data.text, "<Fill/>");
-                       break;
-                default:             //  Illegal
-                       strcpy(obj->data.text, "***Error***");
-                       break;
-              }
-              break;
-    case 0x01:               //  Integer
-              err = createObject(type, length, 0, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              obj->data.integer = readInteger(ptr, length);
-              break;
-    case 0x02:               //  Float
-              err = createObject(type, length, 0, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              obj->data.real = readDouble(ptr, length);
-              break;
-    case 0x03:               //  Date in Elapsed seconds
-              err = createObject(type, length, 0, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              obj->data.date = (time_t)readDouble(ptr, length);
-              break;
-    case 0x04:               //  binary data
-              err = createObject(type, length, length, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              for (i=0; i < length; i++)
-                obj->data.binary[i] = *(ptr++);
-              break;
-    case 0x05:               //  ASCII string
-              err = createObject(type, length, length, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              for (i=0; i < length; i++)
-                obj->data.text[i] = (char)*(ptr++);
-              obj->data.text[length] = '\0';
-              break;
-    case 0x06:               //  UTF16 string
-              err = createObject(type, length, 2 * length, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              for (i=0; i < length; i++) {
-                short int d = *(ptr++);
-                obj->data.utf16[i] = (short int)((d << 8) | *(ptr++));
-              }
-              obj->data.utf16[length] = '\0';
-              break;
-    case 0x08:               //  UID
-              err = createObject(type, length, 0, &obj);
-              if (err != ERROR_NONE)
-                return err;
-              obj->data.uid = readUid(ptr, length);
-              break;
-    case 0x0A:               //  Array
-              err = readArray(cfg, ptr, type, length, &obj);
-              break;
-    case 0x0D:               //  Dictionary
-              err = readDictionary(cfg, ptr, type, length, &obj);
-              break;
-    default:
-              fprintf(stderr, "Unknown object type: %d\n", type);
-              err = createObject(type, length, 0, &obj);
-              break;
-  }
-  if (err == ERROR_NONE)
-    *ptr2 = obj;
-  return err;
-}
-
-void displayHex(CONFIG *cfg, int data)
-{
-  static int in = 0;
-  static char buffer[76];
-  static long offset = 0L;
-  static char hex[] = "0123456789ABCDEF";
-  char *tmp;
-
-  //  If -ve then flush buffer
-  if (data < 0) {
-    if (in > 0) {
-      tmp = buffer + in * 3 + 8 + 2 * (in / 8);
-      while (tmp < buffer+59)
-        *(tmp++) = ' ';
-      buffer[in+59] = '\0';
-      printWithIndent(cfg, buffer, 1);
-    }
-    offset = 0L;
-    in = 0;
-    return;
-  }
-
-  //  If start of line add offset
-  if (in == 0) {
-    sprintf((char *)buffer, "%06lX  ", offset);
-    offset += 16;
-    buffer[8] = '\0';
-  }
-
-  //  Add Byte in hex
-  tmp = buffer + in * 3 + 8 + 2 * (in / 8);
-  *(tmp++) = hex[(data >> 4) & 0x0F];
-  *(tmp++) = hex[data & 0x0F];
-  *(tmp++) = ' ';
-
-  //  Add to character buffer
-  buffer[in+59] = (char)data;
-
-  //  Check if midpoint of buffer
-  if (++in == 8) {
-    *(tmp++) = '-';
-    *(tmp++) = ' ';
-  }
-
-  //  Check if buffer full
-  if (in == 16) {
-    buffer[75] = '\0';
-    printWithIndent(cfg, buffer, 1);
-    in = 0;
-  }
-  return;
-}
-
-void displayObject(CONFIG *cfg, OBJECT *obj, int raw)
-{
-  char text[32];
-  switch (obj->type) {
-    case 0x00:                 //  Singleton
-              printWithIndent(cfg, obj->data.text, 1);
-              break;
-    case 0x01:                 //  Integer
-              if (raw == 0)
-                printWithIndent(cfg, "<number>", 0);
-              sprintf(text, "%ld", obj->data.integer);
-              outputText(cfg, text);
-              if (raw == 0)
-                outputText(cfg, "</number>\n");
-              break;
-    case 0x02:                 //  Float
-              if (raw == 0)
-                printWithIndent(cfg, "<float>", 0);
-              sprintf(text, "%f", obj->data.real);
-              outputText(cfg, text);
-              if (raw == 0)
-                outputText(cfg, "</float>\n");
-              break;
-    case 0x03:                 //  Date
-              if (raw == 0)
-                printWithIndent(cfg, "<date>", 0);
-              outputText(cfg, ctime(&(obj->data.date)));
-              if (raw == 0)
-                outputText(cfg, "</date>\n");
-              break;
-    case 0x04:                 //  Binary data
-              if (raw == 0)
-                printWithIndent(cfg, "<data>", 1);
-              indent += INCREMENT;
-              for (int i=0; i < obj->length; i++)
-                displayHex(cfg, obj->data.binary[i]);
-              displayHex(cfg, -1);
-              indent -= INCREMENT;
-              if (raw == 0)
-                printWithIndent(cfg, "</data>", 1);
-              break;
-    case 0x05:                 //  String
-              if (raw == 0)
-                printWithIndent(cfg, "<string>", 0);
-              outputText(cfg, (obj->data.text));
-              if (raw == 0)
-                outputText(cfg, "</string>\n");
-              break;
-    case 0x06:                 //  UTF16 string
-              if (raw == 0)
-                printWithIndent(cfg, "<utf16>", 1);
-              indent += INCREMENT;
-              for (int i=0; i < obj->length; i++)
-                displayHex(cfg, obj->data.binary[i]);
-              displayHex(cfg, -1);
-              indent -= INCREMENT;
-              if (raw == 0)
-                printWithIndent(cfg, "</utf16>", 1);
-              break;
-    case 0x08:                 //  UID
-              if (raw == 0)
-                printWithIndent(cfg, "<uid>", 0);
-              sprintf(text, "%lu", obj->data.uid);
-              outputText(cfg, text);
-              if (raw == 0)
-                outputText(cfg, "</uid>\n");
-              break;
-    case 0x0A:                 //  Array
-              printWithIndent(cfg, "<array>", 1);
-              indent += INCREMENT;
-              for (int i=0; i < obj->length; i++)
-                displayObject(cfg, obj->data.objects[i], 0);
-              indent -= INCREMENT;
-              printWithIndent(cfg, "</array>", 1);
-              break;
-    case 0x0D:                 //  Dictionary
-              printWithIndent(cfg, "<dict>", 1);
-              indent += INCREMENT;
-              for (int i=0; i < obj->length; i++) {
-                printWithIndent(cfg, "<key>", 0);
-                displayObject(cfg, obj->data.keys[i].key, 1);
-                outputText(cfg, "</key>\n");
-                displayObject(cfg, obj->data.keys[i].value, 0);
-              }
-              indent -= INCREMENT;
-              printWithIndent(cfg, "</dict>", 1);
-              break;
-    default:
-              sprintf(text, "Cannot display type: %d\n", obj->type);
-              outputText(cfg, text);
-              break;
-  }
-  return;
-}
-
-int releaseObject(OBJECT *obj)
-{
-  int i;
-  switch (obj->type) {
-    case 0x00:                        //  Singleton
-    case 0x01:                        //  Int
-    case 0x02:                        //  Float
-    case 0x03:                        //  Date
-    case 0x04:                        //  Binary
-    case 0x05:                        //  ASCII
-    case 0x06:                        //  UTF16
-    case 0x08:                        //  UID
-              break;
-    case 0x0A:                        //  Array
-              for (i=0; i < obj->length; i++)
-                releaseObject(obj->data.objects[i]);
-              break;
-    case 0x0D:                        //  Dictionary
-              for (i=0; i < obj->length; i++) {
-                releaseObject(obj->data.keys[i].key);
-		releaseObject(obj->data.keys[i].value);
-              }
-              break;
-  }
-  free(obj);
-  return ERROR_NONE;
-}
-
-QByteArray parsePlist(QByteArray data)
-{
-  CONFIG cfg;
-  OBJECT *obj;
-  int err = ERROR_NONE;
-  int version;
-  long offset;
-  long length;
-  char text[32];
-  QByteArray result;
-
-  //  Determine the file size and save
-  cfg.bufferLength = data.length();
-  cfg.buffer = (unsigned char *)data.data();
-
-  //  Preset the output buffer
-  cfg.outputBufferLength = 0;
-  cfg.outputBufferIn  = 0;
-  cfg.outputBuffer = NULL;
-
-  //  Read the header
-  err = readHeader(&cfg);
-  if (err != ERROR_NONE)
-    return NULL;  //  err;
-  version = (int)(cfg.buffer[7] - '0');
-
-  //  Read the trailer
-  err = readTrailer(&cfg);
-  if (err != ERROR_NONE)
-    return NULL; //  ERROR_INVALID_TRAILER;
-
-  //  Locate and read the root object
-  offset = cfg.offsetTable[cfg.rootObjectReference];
-  err = readObject(&cfg, offset, &obj);
-
-  //  If no error display the root object and hence the whole object tree
-  if (err != ERROR_NONE)
-    return NULL;
-
-  sprintf(text, "<plist version = \"1.%d\">\n", version);
-  outputText(&cfg, text);
-  displayObject(&cfg, obj, 0);
-  outputText(&cfg, "</plist>\n");
-
-  //  Create return data
-  length = strlen((const char *)(cfg.outputBuffer));
-  result.resize(length);
-  for (int i=0; i < length; i++)
-    result[i] = cfg.outputBuffer[i];
-
-  //  Release assigned memory
-  releaseObject(obj);
-  free(cfg.offsetTable);
-  free(cfg.outputBuffer);
-
-  return result;
-}
-
-//  End of extra data
-
 EditDialog::EditDialog(QWidget* parent)
     : QDialog(parent),
       ui(new Ui::EditDialog),
@@ -714,9 +129,6 @@ void EditDialog::loadData(const QByteArray& data)
         ui->editorText->setEnabled(true);
         sciEdit->setEnabled(true);
         switch (editMode) {
-        case Base64Editor:
-            // No conversion is required
-            // No 'break' is intentional
         case TextEditor:
             // The text widget buffer is now the main data source
             dataSource = TextBuffer;
@@ -772,10 +184,6 @@ void EditDialog::loadData(const QByteArray& data)
         case XmlEditor:
             setDataInBuffer(data, SciBuffer);
             break;
-        case Base64Editor:
-            data2 = QByteArray::fromBase64(data);
-            setDataInBuffer(data2, TextBuffer);
-            break;
         case HexEditor:
             setDataInBuffer(data, HexBuffer);
             break;
@@ -788,34 +196,6 @@ void EditDialog::loadData(const QByteArray& data)
             // Load the text into the text editor
             setDataInBuffer(data, TextBuffer);
 
-            break;
-        }
-        break;
-
-    case Plist:
-        //  Convert data to XML format
-        data2 = parsePlist(data);
-
-        switch (editMode) {
-        case TextEditor:
-        case Base64Editor:
-            setDataInBuffer(data2, TextBuffer);
-            break;
-        case JsonEditor:
-        case XmlEditor:
-            setDataInBuffer(data, SciBuffer);
-            break;
-        case HexEditor:
-            setDataInBuffer(data, HexBuffer);
-            break;
-        case ImageViewer:
-            // The image viewer cannot hold data nor display text.
-
-            // Clear any image from the image viewing widget
-            ui->editorImage->setPixmap(QPixmap(0,0));
-
-            // Load the text into the text editor
-            setDataInBuffer(data2, TextBuffer);
             break;
         }
         break;
@@ -915,11 +295,6 @@ void EditDialog::loadData(const QByteArray& data)
         case ImageViewer:
             // Clear any image from the image viewing widget
             ui->editorImage->setPixmap(QPixmap(0,0));
-            break;
-
-        case Base64Editor:
-            data2 = QByteArray::fromBase64(data);
-            setDataInBuffer(data2, TextBuffer);
             break;
 
         }
@@ -1312,11 +687,6 @@ void EditDialog::editModeChanged(int newMode)
             setDataInBuffer(removedBom + ui->editorText->toPlainText().toUtf8(), HexBuffer);
             break;
 
-        case Base64Editor:
-            data2 = QByteArray::fromBase64(ui->editorText->toPlainText().toUtf8());
-            setDataInBuffer(data2, TextBuffer);
-            break;
-
         case ImageViewer:
             // Clear any image from the image viewing widget
             ui->editorImage->setPixmap(QPixmap(0,0));
@@ -1441,11 +811,6 @@ int EditDialog::checkDataType(const QByteArray& data)
             return JSON;
         else
             return Text;
-    }
-
-    //  Check if it is a 'PLIST' cell (starts "bplist\0\0")
-    if ((data[0] == 'b') && (data[1] == 'p')) {
-      return Plist;
     }
 
     // It's none of the above, so treat it as general binary data
