+++ src/EditDialog.h
@@ -74,8 +74,7 @@ private:
         Null,
         Text,
         JSON,
-        SVG,
-        Plist
+        SVG
     };
 
     // Edit modes and editor stack (this must be aligned with the UI)
@@ -85,8 +84,7 @@ private:
         HexEditor = 1,
         ImageViewer = 2,
         JsonEditor, SciEditor = 3,
-        XmlEditor = 4,
-        Base64Editor = 5
+        XmlEditor = 4
     };
 
     int checkDataType(const QByteArray& data);
