+++ src/extensions/extension-plist.def
@@ -1,0 +1,5 @@
+EXPORTS
+    RegisterExtensionPlist
+    sqlite3_api
+    sqlite3_sqliteplist_init
+    free
